+++ packages/fighting-design/popup/src/props.ts
@@ -27,7 +27,7 @@ export const Props = {
    *
    * @see z-index https://developer.mozilla.org/zh-CN/docs/Web/CSS/z-index
    */
-  zIndex: setNumberProp(1999),
+  zIndex: setNumberProp(),
   /** 自定义遮罩层背景色 */
   maskBackground: setStringProp(),
   /** 遮罩层透明度 */