+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - 修复 `f-button` 组件部分参数响应式丢失问题
+- 修复 `f-drawer` 组件内容部分超出不显示滚动条问题
+- 修复 `f-popup` 组件内容部分超出不显示滚动条问题
+
+- `f-popup` 组件 `z-index` 配置项默认值改为 null
 
 ## 0.40.3 (2023-05-22)